Fix misnamed method, missing JavaDoc details
Upgrading a couple of Maven plugins made the JavaDoc task more stringent
about failing when descriptions had not been provided, or when a method
claims to throw an exception which is not actually thrown, so I added the
missing descriptions and removed the incorrect @throws declaration.

In so doing, I noticed that the method name removeNotificationListener()
in CoreMidiDeviceProvider was previously removedNotificationListener(),
which was inaccurate.