Define a stable automatic module name.
@brunchboy is starting to explore modularizing his smaller projects
and discovered this is often impossible because few of their
dependencies even establish a module name to be used when converted to
an automatic module. See [this
article](http://branchandbound.net/blog/java/2017/12/automatic-module-name/)
for a very good explanation of why this is necessary.

We may want to actually modularize CoreMidi4J someday (not before we
establish Java 9 as a minimum runtime version), but in the mean time,
to be polite to other modular projects that would like to depend on
us, we should declare what module name we intend to use. This change
specifies `uk.co.xfactorylibrarians.coremidi4j` as our intended module
name.