Offer MIDI environment change callbacks on non-mac platforms
It has become clear that the ability to automatically update the user
interface when MIDI devices are added or removed on the Mac offers a
great user experience that I miss when testing my projects on other
platforms. Relocating some delegated methods and adding an optional
daemon thread allows us to provide that feature via simple polling
on platforms where we can't rely on CoreMIDI to do the work for us.